Tighter language in Section 1, more definitions
This commit provides definitions for quorum and various forms of
consensus.  It also explicitly points to a specific version of the
Operating Agreement document as canonical for future reference.

There is also language to prevent a potential loophole in the diversity
clause around corporate term.  The intent is to prevent white cis men
from being able to remove that nuclear scenario in the event that they
become a majority.  The reason for allowing lower limits for updating
the rest of the operating agreement is so that the organization can
more flexibly correct process burdans.