Removed reference to Slack
Slack is fast paced and noisy, and though it will absolutely play
a role in our company, it shouldn't be dictated in our bylaws.

Previously the only time Slack was mentioned as a mechanism for
deliberation was for social media postings.  Frankly we should put
guidelines about social as a social.md file or whatever we want
to call the role responsible for managing our social media.