December 2017 Bylaw Revision
These are mostly clarifying changes, the only meaningful change is
that Quorum has been lowered to 1/2 instead of 2/3.  This must be
put to vote by the sitting overlords.  Technically 100% of overlords
need to approve it since it involves changes to section 0 and 1.

Many thanks to @erekalper for making these corrections... IN A GOOGLE
DOC.