Change minion to Accomplice
Calling the people who do the buildling "minions" is undeniably
hilarious... but also a maybe just a TINY bit disrespectful.  I guess.

Anyway, we think we've found the exact right word: Accomplice.

These proposed bylaw changes update the language, but also expand the
kind of folks who can be considered accomplices.  People can contribute
to a project without writing a line of code, and so we want the project
owners to be able to indicate who they feel are accomplices.

Issue #83 2018 Q4 Serious Bylaw Improvements