Document the BIFFUD tithe
We have to pay 0.4% tax on gross income (before expenses are
subtracted).  This means we need to retain that amount from projects
involving money, so that we don't incur a liability as an organization
that does not have an associated asset.

Furthermore, in order to operate we do need some way of earning some
money.  Since most of our projects are operating essentially at cost, we
decided to institute a policy where 2% of gross income is retained by
the company and put into general operating funds.

Issue #83 2018 Q4 Serious Bylaw Improvements