Codify project petty cash
We want our project approvals to mean something.  Right now there is no
real cost to approving BIFFUD projects, which means there is no real
incentive to reject a project (and no expectation of completion).  The
result is that we have a bunch of approved projects that have gone
nowhere.

This clause should fix that up.

Issue #83 2018 Q4 Serious Bylaw Improvements