Modify profit distribution rules
The previous rules were a little harsh on project owners; saying that
they only get to keep 60% of their profits.  This increases that amount
to 80%, and removes the requirement to distribute money to the corporate
overlords.

Ultimately this change will allow successful projects to have more money
to spend on themselves (possibly re-investing, possibly compensating the
project owners) and also gives the Hive Mind more control over how it
wants to use company assets.