Clarify resources clause
The previous language made it sound like maybe the project owner(s) were
personally paying for things out of pocket.  While they're welcome to do
that... it would have nothing to do with BIFFUD!  This attempts to
clarify the intent, which is that if project money is used to pay for a
consumable resource, that project will get dibs on the resource.