Added upper limit to prelim expense process
This doesn't actually address the question of "what happens if a prelim
expense is approved but then isn't ultimately approved" but it does
narrow the scope of risk.

I think that if a prelim cost is approved... it's bacially approved, and
the protection the corporation has from abuse is the ability to vote
people in / out of their positions.