Revise IP approval clause for GPL-compatible licenses
Our current bylaws specify that BIFFUD projects must use the Apache Software License Version 2.0, with the requirement that any alternative licenses must be approved by a majority of Corporate Overlords. Hot take: the spirit of this clause is to encourage us to do open source work, rather than to make a political statement about which open source license is the “most correct” one. Yet a strict interpretation of this bylaw would be that there is only One True License, and introduce an additional approval requirement for projects with other permissive open source licenses that are not specifically the Apache License Version 2.0, placing these other permissive licenses under the same level of scrutiny as other non-permissive (e.g. “copyleft” licenses).

This revision removes that approval requirement where the project has license imposed by a third-party, if the license is one that is recognized by the open-source community as a permissive open-source license (“GPL-compatible”).

I preserve the intent for BIFFUD projects to adopt the standard “default” license (Apache License Version 2.0), rather than allow projects owners to choose a-la-carte.