Add support for operating expense budget / tax
Over the ... years? a lot of time and energy has been spent getting our
organization to the point that we can do real work.  I don't really know
why that time and energy was spent.  Why are people hiring a company
called Bad Idea Factory.  It makes no sense.

Anyway, they do, and that means we need to actually do things like:

1. Invoice people
2. Pay people who invoice us
3. Make sure we get paid
4. Keep track of which projects have money to spend
5. Review contracts

This stuff takes time and it's not really sustainable to have people do
it for free.  As a result we need to start retaining a reasonable
overhead percentage of funded projects.  No percentage is explicitly set
in the bylaws because it should grow and shrink as need and income
demand.

Note: at the plotting session where we discuss this I will be suggesting
2.6% so that the total BIFFUD tithe + tax will be a nice round 5%