Revamped the money section
This changes a whole lot of stuff.  First, it sets up a structure for
distribution of money that is inline with how we have functioned the
past two years.  By codifying it we are able to pay Corporate Overlords
through targeted allocations (which has significantly better tax
implications than guaranteed payments).

It also makes the Comptroller's role a bit more explicit, provides
restrictions on a project's ability to create debt, creates a structure
for projects to loan (or give) each other money, and makes the need for
invoices more explicit.

This also creates the concept of the BIFFUD Operating Branch which is
responsible for setting up operating processes.

This also adds a bit more clairty around the difference between
a Project's "coffers" and corporate "coffers" while also changing the
definition of Spoils of War to isolate project expense / income
completely from corporate expenses.