Add the concept of COWS and COINS
This makes it so someone can be a Corporate Overlord without being
a legal owner.  This is important for folks who aren't based in the USA
or who otherwise might not want legal ownership of the company for
whatever weird reason.

This is a change to section 0 and 1, so we need complete consensus.