Add bylaws 🔗 to project-application template
The issue template used for project applications referred to the
existence of BIFFUD's corporate bylaws, but didn't link to them.

This might have been intentional, to require people to search for
the bylaws, but the bylaws links contained in this repository point
to the wrong filename (to be fixed in a subsequent commit) or refer
to `documents/operating.md` without naming it as the Bylaws. It's
a twisty puzzle for would-be contributors.

https://gifs.benlk.com/twisty.gif