Correct filename: OPERATING.md is operating.md
This is the only place in the repo where the bylaws link is broken
by having the wrong case. So I'm fixing it. See additional comments
in the 8845375d8f2841aae3e2c835c9a24709e8862b36 commit message as to
why Bad Idea Factory might not want to merge this commit.